pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-analyze-agent"
version = "0.1.0"
description = "Data analysis agent MVP with LangGraph orchestration and local LLM integration."
requires-python = ">=3.10"
dependencies = [
  "langgraph>=0.2.0",
  "langchain>=0.2.0",
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "scikit-learn>=1.3.0",
  "statsmodels>=0.14.0",
  "matplotlib>=3.7.0",
  "seaborn>=0.13.0",
  "pydantic>=2.6.0",
  "pyyaml>=6.0.0",
  "python-dotenv>=1.0.0",
  "joblib>=1.3.0",
  "streamlit>=1.36.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
